Move admin 'view enrollment' a11y tests into their own files
Issue #518 Implement accessibility checking in CI
